Moved ItemContent struct to Content/ItemContent.swift file
